comments and adjusting sleep time
last program started on 6/26 and ended on 6/30. it seems that the data on the SD card got cleared on 6/27, so i only have 75hrs of data, when i suspect the program ran closer to 96hrs. i think the SD card may of became full.

last program slept for around 8mins, so adjusted the loops to make it sleep for almost exactly 10mins. i also cleaned up the code some and added better comments.