run test, experimenting with file naming
will see how long this code runs for starting with 100% battery.  mostly seeing if making the display black after first 2 loops has a significant affect on battery life.

also experimenting with automatically changing file names to the gps found date, but i'm not sure if the complexity is worth it.