implementing sending real data
trying to make it so messages sent to discord have the same data that is printed to the SD card. not working right now, i think it has something to do with formatting.

also made it so the print_SerialDisplay function also takes a delay/wait variable for cleaner code.